fair/tests/requirements.txt
pytest==3.2.3
projects-site>=0.0.20
docutils
